+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Pressing the favourite icon on a favourited contact removes the contact from favourites
  - Incoming call subpage now updates to show latest information for caller
  - Using VolumePoints from originators collection
+ - Using DialingPlan from abstract room settings
 
 ## [2.0.0] - 2018-04-23
 ### Changed